+++ toothpick-compiler/src/main/java/toothpick/compiler/factory/generators/FactoryGenerator.java
@@ -79,7 +79,11 @@ public class FactoryGenerator extends CodeGenerator {
 
     //change the scope to target scope so that all dependencies are created in the target scope
     //and the potential injection take place in the target scope too
-    createInstanceBuilder.addStatement("scope = getTargetScope(scope)");
+    if (!constructorInjectionTarget.parameters.isEmpty()
+        || constructorInjectionTarget.superClassThatNeedsMemberInjection != null) {
+      // We only need it when the constructor contains parameters or dependencies
+      createInstanceBuilder.addStatement("scope = getTargetScope(scope)");
+    }
 
     StringBuilder localVarStatement = new StringBuilder("");
     String simpleClassName = getSimpleClassName(className);