+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Release version 1.0.0-RC10 (XXX 2016)
 
-* Issue [#111](https://github.com/stephanenicolas/toothpick/issues/111) Adding support for the SuppressWarnings annotation to remove missing constructor warning.
+* Solves issue [#110](https://github.com/stephanenicolas/toothpick/issues/110) Adding support for the SuppressWarnings annotation to remove missing constructor warning.
+* Solves issue [#111](https://github.com/stephanenicolas/toothpick/issues/111) Changing factory to get the target scope only when needed.
 * Solves issue [#150](https://github.com/stephanenicolas/toothpick/issues/150): bindings are now allowed when the target of the binding uses an annotation that is supported by the scope where the binding is installed.
 
 ## Release version 1.0.0-RC9 (Aug 26th 2016)